move buffer from iterable back
